requirements.txt
anthropic
libtmux
openai
python-dotenv
